Move pool owner actions from actions.ts
